Move types from adapter to types package
